+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## master
 [v5.0.0-beta.2...master](https://github.com/deployphp/deployer/compare/v5.0.0-beta.2...master)
 
--
+### Fixed
+- Fixed ssh multiplexing master connection initializing
 
 ## v5.0.0-beta.3
 [v5.0.0-beta.2...v5.0.0-beta.3](https://github.com/deployphp/deployer/compare/v5.0.0-beta.2...v5.0.0-beta.3)